Add gimpcolorspace object.
2000-01-04  Tor Lillqvist  <tml@iki.fi>

* libgimp/makefile.{cygwin.msc}: Add gimpcolorspace object.

* libgimp/gimp.def: Add functions from it.

Fixes from Hans Breuer:

* app/datafiles.c: redefine the executable flag for Win32
to _S_IREAD, to get _all_ files from the plug-in dirs as
executables (including scripts)

* app/main.c: Win32-specific changes to allow building Gimp as a
console application, with all its benefits (like inheriting the
console), but hide it if the user doesn't want it. Also, if stdout
goes to a console, give the user a chance to read the help or
version messages. (tml: I am not convinced that it is better to
build gimp as a console application, but let's try it this way for
a while.)

* app/makefile.{cygwin,msc}: Build as console application, and
link with shell32 library.

* app/paint_core.c (paint_core_motion): Pass the value of a call
to the function gimage_active_drawable() to the paint_func,
instead of just passing the address of gimage_active_drawable...

(tml: This code is only called when the TOOL_TRACES_ON_WINDOW flag
is on, and only the clone tool sets that, and the clone tool's
paint_func doesn't use the drawable argument, so this hasn't
caused any trouble.)

* app/plug_in.c: On Win32, to support scripts, use new function
xspawn() instead of _spawnv. Add some more code to properly kill
plug-ins.

* libgimp/color_display.h: Add G_MODULE_EXPORT declarations.

+++ app/plug-in/plug-in-message.c
@@ -802,6 +802,59 @@ plug_in_destroy (PlugIn *plug_in)
     }
 }
 
+#ifdef G_OS_WIN32
+/* The Microsoft _spawnv() does not allow to run scripts. But
+ * this is essential to get scripting extension up and running.
+ * Following the replacement function xspawnv().
+ */
+int
+xspawnv (int                mode,
+	 const char        *cmdname,
+	 const char *const *argv )
+{
+  char sExecutable[_MAX_PATH*2];
+  char** sArgsList;
+  char sCmndLine[1024];
+  char* sPath;
+  HINSTANCE hInst;
+  int i;
+  int pid;
+
+  /* only use it if _spawnv fails */
+  pid = _spawnv(mode, cmdname, argv);
+  if (pid != -1) return pid;
+
+  /* stuff parameters into one cmndline */
+  sCmndLine[0] = 0;
+  for (i = 1; argv[i] != NULL; i++)
+    {
+       strcat(sCmndLine, argv[i]);
+       strcat(sCmndLine, " ");
+    }
+  /* remove last blank */
+  sCmndLine[strlen(sCmndLine)-1] = 0;
+
+  /* do we really need _spawnv (ShelExecute seems not to do it)*/
+  if (32 <= (int)FindExecutable (cmdname, 
+				 gimp_directory (),
+				 sExecutable))
+    {
+      //g_print("_spawnlp %s %s %s", sExecutable, cmdname, sCmndLine);
+      
+      pid = _spawnlp(mode, sExecutable, "-c", cmdname, sCmndLine, NULL);
+    }
+  else
+    {
+      g_warning("Execution error for: %s", cmdname);
+      return -1;
+    }
+  return pid;
+}
+#undef _spawnv
+#define _spawnv xspawnv
+
+#endif /* G_OS_WIN32 */
+
 gint
 plug_in_open (PlugIn *plug_in)
 {
@@ -989,7 +1042,26 @@ plug_in_close (PlugIn *plug_in,
         waitpid (plug_in->pid, &status, 0);
 #else
       if (kill_it && plug_in->pid)
-	TerminateProcess ((HANDLE) plug_in->pid, 0);
+	{
+	  /* Trying to avoid TerminateProcess (does mostly work).
+	   * Otherwise some of our needed DLLs may get into an unstable state
+	   * (see Win32 API docs).
+	   */
+	  DWORD dwExitCode = STILL_ACTIVE;
+	  DWORD dwTries  = 10;
+	  while ((STILL_ACTIVE == dwExitCode)
+		 && GetExitCodeProcess((HANDLE) plug_in->pid, &dwExitCode)
+		 && (dwTries > 0))
+	    {
+	      Sleep(10);
+	      dwTries--;
+	    }
+	  if (STILL_ACTIVE == dwExitCode)
+	    {
+	      g_warning("Terminating %s ...", plug_in->args[0]);
+	      TerminateProcess ((HANDLE) plug_in->pid, 0);
+	    }
+	}
 #endif
 
       /* Remove the input handler.
